deploy_real/onboard/requirements.txt
pyrealsense2
pyzmq<27
numpy
opencv-python-headless
